// src/app/models/userview.ts
export class userview{
    user_id: any;
    first_name: any;
    last_name: any;
    email: any;
    mobileno: any;
    city: any;
    state: any;
    postcode: any;
    country: any;
    notification: any;
    door:  any;
    profile_pic: any;
    response:any;
    petlist:any;
    productcompany:any;
}